# openapi/phase1.yaml
openapi: 3.0.3
info:
  title: HARP Comptabilité V3 API - Phase 1
  description: |
    API de gestion comptable pour HARP E-commerce.
    
    ## Authentification
    Toutes les routes nécessitent une session admin valide (NextAuth.js).
    Cookie de session requis pour toutes les requêtes.
    
    ## Phase 1 - Fonctionnalités
    - Gestion des achats (création, réception)
    - Gestion de l'inventaire (CRUD, ajustements)
    - Calcul CUMP automatique
    - Réconciliation des stocks
  version: 3.0.0-phase1
  contact:
    name: HARP Team
    email: dev@harp.dz
servers:
  - url: http://localhost:3000
    description: Development server
...
paths:
...
  /api/v3/compta/purchases:
    get:
      tags: [Purchases]
      summary: Liste des achats
      description: Retourne une liste paginée des achats avec filtres optionnels
      operationId: listPurchases
      security:
        - cookieAuth: []
      parameters:
        - $ref: '#/components/parameters/page'
        - $ref: '#/components/parameters/pageSize'
        - name: status
          in: query
          schema:
            type: string
            enum: [DRAFT, ORDERED, PARTIAL, RECEIVED, CANCELLED]
        - name: supplierId
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Liste des achats
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PurchaseListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags: [Purchases]
      summary: Créer un achat
      description: Crée un nouvel achat en statut DRAFT
      operationId: createPurchase
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PurchaseCreateInput'
            example:
              supplierId: "clp123abc"
              invoiceNumber: "FAC-2025-001"
              items:
                - inventoryItemId: "inv001"
                  quantityOrdered: 100
                  unitPrice: 600.00
      responses:
        '201':
          description: Achat créé
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PurchaseDetailResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: next-auth.session-token
  parameters:
    page:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    pageSize:
      name: pageSize
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
...
  responses:
    Unauthorized:
      description: Non authentifié
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error: "Non authentifié"
            code: "NOT_AUTHENTICATED"
    ValidationError:
      description: Erreur de validation
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ValidationErrorResponse'
...
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
        code:
          type: string
    ValidationErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
          example: "Validation échouée"
        details:
          type: object
          additionalProperties:
            type: array
            items:
              type: string
    PaginationMeta:
      type: object
      properties:
        page:
          type: integer
        pageSize:
          type: integer
        total:
          type: integer
        totalPages:
          type: integer
...
    PurchaseCreateInput:
      type: object
      required: [supplierId, items]
      properties:
        supplierId:
          type: string
        invoiceNumber:
          type: string
        invoiceDate:
          type: string
          format: date-time
        expectedDate:
          type: string
          format: date-time
        notes:
          type: string
        items:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/PurchaseItemInput'
    PurchaseItemInput:
      type: object
      required: [inventoryItemId, quantityOrdered, unitPrice]
      properties:
        inventoryItemId:
          type: string
        quantityOrdered:
          type: number
          minimum: 0
        unitPrice:
          type: number
          minimum: 0
...
    PurchaseListResponse:
      type: object
      properties:
        success:
          type: boolean
        items:
          type: array
          items:
            $ref: '#/components/schemas/PurchaseSummary'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
    PurchaseSummary:
      type: object
      properties:
        id:
          type: string
        purchaseNumber:
          type: string
        status:
          type: string
        totalAmount:
          type: number
        supplier:
          type: object
          properties:
            id:
              type: string
            name:
              type: string
            code:
              type: string
        createdAt:
          type: string
          format: date-time
    PurchaseDetailResponse:
      type: object
      properties:
        success:
          type: boolean
        purchase:
          $ref: '#/components/schemas/PurchaseDetail'
    PurchaseDetail:
      type: object
      properties:
        id:
          type: string
        purchaseNumber:
          type: string
        status:
          type: string
        totalAmount:
          type: number
        supplier:
          type: object
        items:
          type: array
          items:
            $ref: '#/components/schemas/PurchaseItemDetail'
    PurchaseItemDetail:
      type: object
      properties:
        id:
          type: string
        inventoryItem:
          $ref: '#/components/schemas/InventoryItemSummary'
        quantityOrdered:
          type: number
        quantityReceived:
          type: number
        unitPrice:
          type: number
        totalPrice:
          type: number
    InventoryItemSummary:
      type: object
      properties:
        id:
          type: string
        sku:
          type: string
        name:
          type: string
        quantity:
          type: number
        averageCost:
          type: number
...